+++ src/lib/components/forms/form/helper/terra-form.helper.ts
@@ -1,5 +1,5 @@
 import { AbstractControl, FormArray, FormControl, FormGroup, ValidatorFn, Validators } from '@angular/forms';
-import { isArray, isNullOrUndefined, isObject, isString } from 'util';
+import { isNullOrUndefined } from 'util';
 import { StringHelper } from '../../../../helpers/string.helper';
 import { TerraValidators } from '../../../../validators/validators';
 import { TerraFormFieldHelper } from './terra-form-field.helper';
@@ -109,18 +109,18 @@ export class TerraFormHelper {
             let defaultValue: any = TerraFormFieldHelper.parseDefaultValue(formField);
             if (formField.isList) {
                 let formControls: Array<AbstractControl> = [];
-                if (!isNullOrUndefined(values) && isArray(values[formFieldKey])) {
+                if (!isNullOrUndefined(values) && Array.isArray(values[formFieldKey])) {
                     formControls = (values[formFieldKey] as Array<any>).map((value: any, index: number) => {
                         return this.createNewControl(value || defaultValue[index], formField);
                     });
                 }
-                if (isString(formField.isList)) {
+                if (typeof formField.isList === 'string') {
                     this._fitControlsToRange(formField, formControls);
                 }
                 controls[formFieldKey] = new FormArray(formControls, this.generateFormArrayValidators(formField));
             } else if (!isNullOrUndefined(formField.children)) {
                 let value: Object =
-                    !isNullOrUndefined(values) && isObject(values[formFieldKey])
+                    !isNullOrUndefined(values) && typeof values[formFieldKey] === 'object'
                         ? values[formFieldKey]
                         : defaultValue || null;
                 controls[formFieldKey] = this.parseReactiveForm(formField.children, value);
@@ -145,7 +145,7 @@ export class TerraFormHelper {
         formFields: { [key: string]: TerraFormFieldInterface },
         values: any
     ): any {
-        if (form instanceof FormGroup && !isObject(values)) {
+        if (form instanceof FormGroup && !(values !== null && typeof values === 'object')) {
             return;
         }
 
@@ -171,7 +171,8 @@ export class TerraFormHelper {
             } else if (
                 !isNullOrUndefined(formField.children) &&
                 control instanceof FormGroup &&
-                isObject(controlValues)
+                controlValues !== null &&
+                typeof controlValues === 'object'
             ) {
                 values[formControlKey] = this.updateFormArrays(control, formField.children, controlValues);
             }